fixed bug with type cast

+++ lib/util.c
@@ -146,7 +146,7 @@ void fprintw(FILE *f, const char *s, size_t w)
 
 // parse a string representing a number/percentage of targets
 // A percentage is interpreted as percent of the search space, so 2^32 * port_count
-unsigned long long int parse_max_targets(char *max_targets, int port_count)
+uint64_t parse_max_targets(char *max_targets, int port_count)
 {
 	assert(port_count > 0);
 	char *end;
@@ -163,8 +163,8 @@ unsigned long long int parse_max_targets(char *max_targets, int port_count)
 	}
 	if (v <= 0) {
 		return 0;
-	} else if (v >= ((unsigned long long int)1 << 32 * port_count)) {
-		return ((unsigned long long int)1 << 32 * port_count);
+	} else if (v >= (((unsigned long long int)1 << 32) * port_count)) {
+		return (((unsigned long long int)1 << 32) * port_count);
 	} else {
 		return v;
 	}
